Move backend files such that related localstore files are separated from other files
