+++ ledger/participant-integration-api/src/main/scala/platform/localstore/PersistentPartyRecordStore.scala
@@ -25,7 +25,7 @@ import com.daml.platform.localstore.api.{LedgerPartyExists, PartyRecordStore, Pa
 import com.daml.platform.localstore.utils.LocalAnnotationsUtils
 import com.daml.platform.server.api.validation.ResourceAnnotationValidation
 import com.daml.platform.store.DbSupport
-import com.daml.platform.store.backend.PartyRecordStorageBackend
+import com.daml.platform.store.backend.localstore.PartyRecordStorageBackend
 
 import scala.concurrent.{ExecutionContext, Future}
