+++ ledger/participant-integration-api/src/main/scala/platform/store/backend/StorageBackend.scala
@@ -3,7 +3,7 @@
 
 package com.daml.platform.store.backend
 
-import com.daml.ledger.api.domain.{LedgerId, ParticipantId, PartyDetails, UserRight}
+import com.daml.ledger.api.domain.{LedgerId, ParticipantId, PartyDetails}
 import com.daml.ledger.api.v1.command_completion_service.CompletionStreamResponse
 import com.daml.ledger.configuration.Configuration
 import com.daml.ledger.offset.Offset
@@ -21,7 +21,6 @@ import com.daml.platform.{
   PackageId,
   Party,
   TransactionId,
-  UserId,
 }
 import com.daml.platform.store.EventSequentialId
 import com.daml.platform.store.dao.events.Raw
@@ -32,10 +31,8 @@ import com.daml.platform.store.entries.{ConfigurationEntry, PackageLedgerEntry, 
 import com.daml.platform.store.interfaces.LedgerDaoContractsReader.KeyState
 import com.daml.platform.store.interning.StringInterning
 import com.daml.scalautil.NeverEqualsOverride
-
 import java.sql.Connection
 import javax.sql.DataSource
-import com.daml.lf.data.Ref
 
 import scala.annotation.unused
 
@@ -185,40 +182,6 @@ trait PartyStorageBackend {
   )(connection: Connection): Vector[(Offset, PartyLedgerEntry)]
   def parties(parties: Seq[Party])(connection: Connection): List[PartyDetails]
   def knownParties(connection: Connection): List[PartyDetails]
-}
-
-// TODO um-for-hub: Consider extracting from StorageBackend hierarchy
-trait PartyRecordStorageBackend extends ResourceVersionOpsBackend {
-
-  def getPartyRecord(party: Ref.Party)(
-      connection: Connection
-  ): Option[PartyRecordStorageBackend.DbPartyRecord]
-
-  def createPartyRecord(partyRecord: PartyRecordStorageBackend.DbPartyRecordPayload)(
-      connection: Connection
-  ): Int
-
-  def getPartyAnnotations(internalId: Int)(connection: Connection): Map[String, String]
-
-  def addPartyAnnotation(internalId: Int, key: String, value: String, updatedAt: Long)(
-      connection: Connection
-  ): Unit
-
-  def deletePartyAnnotations(internalId: Int)(connection: Connection): Unit
-
-}
-
-object PartyRecordStorageBackend {
-  case class DbPartyRecordPayload(
-      party: Ref.Party,
-      resourceVersion: Long,
-      createdAt: Long,
-  )
-
-  case class DbPartyRecord(
-      internalId: Int,
-      payload: DbPartyRecordPayload,
-  )
 }
 
 trait PackageStorageBackend {
@@ -465,79 +428,6 @@ trait StringInterningStorageBackend {
   def loadStringInterningEntries(fromIdExclusive: Int, untilIdInclusive: Int)(
       connection: Connection
   ): Iterable[(Int, String)]
-}
-
-trait ResourceVersionOpsBackend {
-  def compareAndIncreaseResourceVersion(
-      internalId: Int,
-      expectedResourceVersion: Long,
-  )(connection: Connection): Boolean
-
-  def increaseResourceVersion(
-      internalId: Int
-  )(connection: Connection): Boolean
-}
-trait UserManagementStorageBackend extends ResourceVersionOpsBackend {
-
-  def createUser(user: UserManagementStorageBackend.DbUserPayload)(connection: Connection): Int
-
-  def addUserAnnotation(internalId: Int, key: String, value: String, updatedAt: Long)(
-      connection: Connection
-  ): Unit
-
-  def deleteUserAnnotations(internalId: Int)(connection: Connection): Unit
-
-  def getUserAnnotations(internalId: Int)(connection: Connection): Map[String, String]
-
-  def deleteUser(id: UserId)(connection: Connection): Boolean
-
-  def getUser(id: UserId)(connection: Connection): Option[UserManagementStorageBackend.DbUserWithId]
-
-  def getUsersOrderedById(fromExcl: Option[UserId] = None, maxResults: Int)(
-      connection: Connection
-  ): Vector[UserManagementStorageBackend.DbUserWithId]
-
-  def addUserRight(internalId: Int, right: UserRight, grantedAt: Long)(
-      connection: Connection
-  ): Unit
-
-  /** @return true if the right existed and we have just deleted it.
-    */
-  def deleteUserRight(internalId: Int, right: UserRight)(connection: Connection): Boolean
-
-  def userRightExists(internalId: Int, right: UserRight)(connection: Connection): Boolean
-
-  def getUserRights(internalId: Int)(
-      connection: Connection
-  ): Set[UserManagementStorageBackend.DbUserRight]
-
-  def countUserRights(internalId: Int)(connection: Connection): Int
-
-  def updateUserPrimaryParty(internalId: Int, primaryPartyO: Option[Ref.Party])(
-      connection: Connection
-  ): Boolean
-
-  def updateUserIsDeactivated(
-      internalId: Int,
-      isDeactivated: Boolean,
-  )(connection: Connection): Boolean
-
-}
-
-object UserManagementStorageBackend {
-  case class DbUserPayload(
-      id: Ref.UserId,
-      primaryPartyO: Option[Ref.Party],
-      isDeactivated: Boolean,
-      resourceVersion: Long,
-      createdAt: Long,
-  )
-
-  case class DbUserWithId(
-      internalId: Int,
-      payload: DbUserPayload,
-  )
-  case class DbUserRight(domainRight: UserRight, grantedAt: Long)
 }
 
 trait MeteringStorageReadBackend {
