+++ ledger/participant-integration-api/src/main/scala/platform/store/backend/StorageBackendFactory.scala
@@ -5,6 +5,10 @@ package com.daml.platform.store.backend
 
 import com.daml.platform.store.DbType
 import com.daml.platform.store.backend.h2.H2StorageBackendFactory
+import com.daml.platform.store.backend.localstore.{
+  PartyRecordStorageBackend,
+  UserManagementStorageBackend,
+}
 import com.daml.platform.store.backend.oracle.OracleStorageBackendFactory
 import com.daml.platform.store.backend.postgresql.PostgresStorageBackendFactory
 import com.daml.platform.store.cache.LedgerEndCache
