+++ ledger/participant-integration-api/src/main/scala/platform/store/backend/localstore/ParticipantMetadataBackend.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.store.backend.common
+package com.daml.platform.store.backend.localstore
 
 import java.sql.Connection
