+++ ledger/participant-integration-api/src/main/scala/platform/store/backend/localstore/ResourceVersionOpsBackend.scala
@@ -1,0 +1,17 @@
+// Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.platform.store.backend.localstore
+
+import java.sql.Connection
+
+trait ResourceVersionOpsBackend {
+  def compareAndIncreaseResourceVersion(
+      internalId: Int,
+      expectedResourceVersion: Long,
+  )(connection: Connection): Boolean
+
+  def increaseResourceVersion(
+      internalId: Int
+  )(connection: Connection): Boolean
+}
