+++ ledger/participant-integration-api/src/test/lib/scala/platform/localstore/PersistentUserStoreTests.scala
@@ -8,13 +8,13 @@ import java.sql.Connection
 import com.daml.api.util.TimeProvider
 import com.daml.lf.data.Ref
 import com.daml.metrics.Metrics
-import com.daml.platform.store.backend.{
+import com.daml.platform.store.backend.StorageBackendProvider
+import com.daml.platform.store.backend.localstore.UserManagementStorageBackend.DbUserPayload
+import com.daml.platform.store.backend.localstore.{
   ResourceVersionOpsBackend,
-  StorageBackendProvider,
   UserManagementStorageBackend,
+  UserManagementStorageBackendImpl,
 }
-import com.daml.platform.store.backend.UserManagementStorageBackend.DbUserPayload
-import com.daml.platform.store.backend.common.UserManagementStorageBackendImpl
 import org.scalatest.freespec.AsyncFreeSpec
 
 trait PersistentUserStoreTests
