+++ ledger/participant-integration-api/src/test/lib/scala/platform/store/backend/StorageBackendProvider.scala
@@ -13,8 +13,12 @@ import com.daml.platform.store.interning.MockStringInterning
 import com.daml.testing.oracle.OracleAroundAll
 import com.daml.testing.postgresql.PostgresAroundAll
 import org.scalatest.Suite
-
 import java.sql.Connection
+
+import com.daml.platform.store.backend.localstore.{
+  PartyRecordStorageBackend,
+  UserManagementStorageBackend,
+}
 
 /** Creates a database and a [[TestBackend]].
   * Used by [[StorageBackendSpec]] to run all StorageBackend tests on different databases.
