+++ ledger/participant-integration-api/src/test/lib/scala/platform/store/backend/StorageBackendTestsPartyRecord.scala
@@ -7,6 +7,7 @@ import java.sql.SQLException
 import java.util.UUID
 
 import com.daml.lf.data.Ref
+import com.daml.platform.store.backend.localstore.PartyRecordStorageBackend
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.{Inside, OptionValues}
